backups/2025-12-05_19-02-01/files/requirements.txt
# Sistema de Cotizaciones BMC Uruguay
# Dependencias requeridas

# Módulos estándar de Python (incluidos por defecto)
# json
# datetime
# decimal
# csv
# typing
# dataclasses
# re
# time

# Dependencias externas para funcionalidades avanzadas
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3

# Para generación de PDFs (opcional)
reportlab>=3.6.0

# Para procesamiento de datos (opcional)
pandas>=1.3.0
openpyxl>=3.0.7

# Para interfaz web (opcional)
flask>=2.0.1
jinja2>=3.0.1

# Para testing (opcional)
pytest>=6.2.4
pytest-cov>=2.12.1

# Para sistema de actualización automática
schedule>=1.2.0

# Para integración de modelos de IA (OpenAI, Groq, Gemini, xAI/Grok)
openai>=1.0.0  # Also used for xAI/Grok (OpenAI-compatible API)
groq>=0.4.0
google-genai>=0.2.0  # New recommended SDK for Gemini (2025)
google-generativeai>=0.3.0  # Fallback for old SDK compatibility
python-dotenv>=1.0.0

# Para API FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Para MongoDB
pymongo>=4.5.0

# Para Google Sheets API
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Para n8n API integration
# requests ya está incluido arriba (línea 15)

# Para rate limiting
slowapi>=0.1.9

# Para Vector Database (Qdrant)
qdrant-client>=1.7.0

# Para caching y sesiones (Redis)
redis>=5.0.0

# Para monitoreo de sistema
psutil>=5.9.0

# Para linting y formateo de código Python
ruff>=0.1.0
black>=23.0.0
cryptography>=41.0.0
